+++ src/layouts/AuthFormLayout.tsx
@@ -1,12 +1,12 @@
 import { FormEvent, ReactNode } from 'react';
 
-type AuthFormProps = {
+type AuthFormLayoutProps = {
   children: ReactNode;
   onSubmit: (e: FormEvent<HTMLFormElement>) => void;
   marginTop: 'mt-34.9' | 'mt-40';
 };
 
-export default function AuthForm({ children, onSubmit, marginTop }: AuthFormProps) {
+export default function AuthFormLayout({ children, onSubmit, marginTop }: AuthFormLayoutProps) {
   return (
     <>
       <section className="mt-40 text-large text-main">
